SSAFY/lectures/99_toys/window_sub_screen/requirements.txt
MouseInfo==0.1.3
mss==9.0.2
pillow==10.4.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
screeninfo==0.8.1
